requirements/production.txt
# Install everything that needs to be compiled
-r compiled.txt

# Generyc python stuff
pytz

# Django stuff
Django>=1.8,<1.9
django-timezone-field

# Templates
django_compressor

# Security
bleach
#cef

# Logging
#raven

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Admin
django-debug-toolbar

# Feature testing
#django-waffle

# 3rd party authentication
django-allauth==0.23.0
